requirements.txt
streamlit
opencv-python
google-generativeai
pyttsx3
numpy
pillow
python-dotenv